autorizacion por tokne agreagado a la ap Autenticacion (Logout y eliminar usuario)

+++ Autenticacion/views.py
@@ -1,5 +1,7 @@
 from datetime import datetime
 from django.shortcuts import render
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from .serializer import UserSerializer, UserInfoSerializer
@@ -8,6 +10,7 @@ from django.contrib.auth.models import User
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.authtoken.models import Token
 from django.contrib.sessions.models import Session
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
 
 # Create your views here.
 class Autenticacion(APIView):
@@ -21,16 +24,32 @@ class Autenticacion(APIView):
 
         return Response('Usuario Creado', status=status.HTTP_201_CREATED)
 
-    def delete(self, request, pk):
+
+@api_view(['DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def delete(request, pk):
         # Luego cuando se monte a la base en la nube, cambiarlo por 1
         if pk == '8':
             return Response('No se debe eliminar al admin', status=status.HTTP_401_UNAUTHORIZED)
+
+        authorization_header = request.META.get('HTTP_AUTHORIZATION', None).split()
+        token =  Token.objects.get(key = authorization_header[1]) 
+
+
         user = User.objects.get(id = pk)
+        
+        if user != token.user:
+            return Response('Token perteneciente a otro usuario', status=status.HTTP_401_UNAUTHORIZED)
+
         user.delete()
         return Response('Usuario Eliminado', status=status.HTTP_200_OK)
 
 
+
+
 class Login(ObtainAuthToken):
+
     def post(self, request, *args, **kwargs):
         login_serializer = self.serializer_class(data = request.data, context = {request:request})
         if login_serializer.is_valid():
@@ -65,12 +84,14 @@ class Login(ObtainAuthToken):
 
 
 class Logout(APIView):
-    def post(self, request):
-        data =  request.data
-        print(data)
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
+    
+    def get(self, request):
+        authorization_header = request.META.get('HTTP_AUTHORIZATION', None).split()
+        data =  authorization_header[1]
         try:
-            token = Token.objects.get(key = data['token'])
-            print(token.key)
+            token = Token.objects.get(key = data)
             user = token.user
             all_sessions = Session.objects.filter(expire_date__gte = datetime.now())
             if all_sessions.exists():
